rename package distribution to remote
